+++ Sources/TokenVisitor/SyntaxClassifiable/CodeBlockItemListSyntax.swift
@@ -1,5 +1,5 @@
 //
-//  SyntaxClassifiable.swift
+//  CodeBlockItemListSyntax.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -30,11 +30,10 @@
 import Foundation
 import SwiftSyntax
 
-/// Protocol for syntax nodes that can classify themselves by their semantic role.
-///
-/// This protocol allows syntax nodes to self-identify their classification,
-/// making the classification process more type-safe and extensible.
-package protocol SyntaxClassifiable: Sendable {
-  /// Returns the semantic classification of this syntax node type.
-  static var syntaxType: SyntaxType { get }
+/// Extension for CodeBlockItemListSyntax to conform to SyntaxClassifiable.
+extension CodeBlockItemListSyntax: SyntaxClassifiable {
+  /// Code block item lists are classified as `.collection` type.
+  internal static var syntaxType: SyntaxType {
+    .collection
+  }
 }
